메세지 출력 생성 확인
